Reorganise: move most of kas::data to kas::widget::view and rest to kas::updatable
Rationale: SingleData, ListData etc. are strongly
correlated with the view widgets thus belong there.
FilteredList applies only to ListData.
SharedRc is mostly impls on ListData etc. traits.
