+++ kas-wgpu/examples/sync-counter.rs
@@ -5,10 +5,9 @@
 
 //! A counter synchronised between multiple windows
 
-use kas::data::SharedRc;
 use kas::event::{Manager, Response, VoidMsg};
 use kas::macros::{make_widget, VoidMsg};
-use kas::widget::view::SingleView;
+use kas::widget::view::{SharedRc, SingleView};
 use kas::widget::{TextButton, Window};
 
 #[derive(Clone, Debug, VoidMsg)]
