+++ src/widget/view/list_view.rs
@@ -5,11 +5,11 @@
 
 //! List view widget
 
-use super::{driver, Driver, SelectionMode};
-use kas::data::{ListData, UpdatableAll};
+use super::{driver, Driver, ListData, SelectionMode};
 use kas::event::{ChildMsg, CursorIcon, GrabMode, PressSource};
 use kas::layout::solve_size_rules;
 use kas::prelude::*;
+use kas::updatable::UpdatableAll;
 #[allow(unused)] // doc links
 use kas::widget::ScrollBars;
 use kas::widget::{ScrollComponent, Scrollable};
